Moved frontend project root into src and relocated related settings to project root.
