+++ src/app/src/_mantine.scss
