+++ src/app/src/routes/campaign_detail/page.tsx
@@ -3,15 +3,15 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { Fragment, useContext, useId, useMemo, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import { Anchored } from "src/components/anchored.tsx";
-import { CharacterSelector } from "src/components/character_selector/character_selector.tsx";
-import { UploadModal } from "src/components/upload-modal.tsx";
-import { useCampaign, useCombatsForCampaign, useWatchCampaign } from "src/hooks/api_hooks.ts";
-import { CharacterCard } from "src/components/character_card/card.tsx";
-import { CharacterEditor } from "src/components/character_editor/character_editor.tsx";
-import { createCharacter, createCombat, CreateCombatUpdate, deleteCombat, updateCampaign } from "src/services/api.ts";
-import { ErrorContext } from "src/services/contexts.ts";
-import { Character, Combat } from "src/types/models.ts";
+import { Anchored } from "components/anchored.tsx";
+import { CharacterSelector } from "components/character_selector/character_selector.tsx";
+import { UploadModal } from "components/upload-modal.tsx";
+import { useCampaign, useCombatsForCampaign, useWatchCampaign } from "hooks/api_hooks.ts";
+import { CharacterCard } from "components/character_card/card.tsx";
+import { CharacterEditor } from "components/character_editor/character_editor.tsx";
+import { createCharacter, createCombat, CreateCombatUpdate, deleteCombat, updateCampaign } from "services/api.ts";
+import { ErrorContext } from "services/contexts.ts";
+import { Character, Combat } from "types/models.ts";
 import { ActionIcon, Button, Card, Flex, Modal, Text, Image, Title, Stack, Group, Divider, Box, useModalsStack } from "@mantine/core";
 import Element = React.JSX.Element;
