+++ src/app/src/services/contexts.ts
@@ -1,5 +1,5 @@
 import { Context, createContext, useState } from "react";
-import { CampaignDetails, Character } from "src/types/models.ts";
+import { CampaignDetails, Character } from "types/models.ts";
 
 type ReactStateController<T> = ReturnType<typeof useState<T>>;
