+++ vite.config.ts
@@ -1,8 +1,17 @@
+import * as path from "node:path";
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import tsconfigPaths from "vite-tsconfig-paths";
 
+const fromRoot = (append: string) => path.resolve(__dirname, append)
+
 export default defineConfig({
+  root: fromRoot('src/app'),
+  resolve: {
+    alias: {
+      'src': fromRoot('src/app/src')
+    }
+  },
   plugins: [
     react({
       babel: {
@@ -15,7 +24,7 @@ export default defineConfig({
     tsconfigPaths()
   ],
   build: {
-    outDir: '../build/resources/main/static/app',
+    outDir: fromRoot('build/resources/main/static/app'),
     emptyOutDir: true,
   },
   server: {
